+++ test/conftest.py
@@ -1,12 +1,10 @@
 import pytest
-from os.path import join, exists
-from os import makedirs
+from os.path import join
 from uuid import uuid4
-from shutil import copytree, copyfile
-from PyQt5.QtCore import QTimer, QVariant
+from shutil import copytree
+from PyQt5.QtCore import QTimer
 from PyQt5.QtWidgets import QApplication
-from qgis.core import QgsProject, QgsVectorLayer, QgsField, QgsFeature
-from qgis.core import QgsPointXY, QgsGeometry, QgsCoordinateReferenceSystem
+from qgis.core import QgsProject
 from qaequilibrae.qaequilibrae import AequilibraEMenu
 from qaequilibrae.modules.common_tools import ReportDialog
 
@@ -21,6 +19,7 @@ def ae(qgis_iface) -> AequilibraEMenu:
     ae = AequilibraEMenu(qgis_iface)
     yield ae
     qgis_iface.messageBar().messages = {0: [], 1: [], 2: [], 3: []}
+    QgsProject.instance().removeAllMapLayers()
 
 
 @pytest.fixture(scope="function")
@@ -33,6 +32,7 @@ def ae_with_project(qgis_iface, folder_path) -> AequilibraEMenu:
     yield ae
     ae.run_close_project()
     qgis_iface.messageBar().messages = {0: [], 1: [], 2: [], 3: []}
+    QgsProject.instance().removeAllMapLayers()
 
 
 @pytest.fixture(scope="function")
@@ -66,6 +66,7 @@ def pt_project(qgis_iface, folder_path) -> AequilibraEMenu:
     yield ae
     ae.run_close_project()
     qgis_iface.messageBar().messages = {0: [], 1: [], 2: [], 3: []}
+    QgsProject.instance().removeAllMapLayers()
 
 
 @pytest.fixture(scope="function")
@@ -78,39 +79,7 @@ def pt_no_feed(qgis_iface, folder_path) -> AequilibraEMenu:
     yield ae
     ae.run_close_project()
     qgis_iface.messageBar().messages = {0: [], 1: [], 2: [], 3: []}
-
-
-@pytest.fixture(scope="function")
-def load_layers_from_csv():
-    import csv
-
-    path_to_csv = "test/data/SiouxFalls_project/SiouxFalls_od.csv"
-    datalayer = QgsVectorLayer("None?delimiter=,", "open_layer", "memory")
-
-    fields = [
-        QgsField("O", QVariant.Int),
-        QgsField("D", QVariant.Int),
-        QgsField("Ton", QVariant.Double),
-    ]
-    datalayer.dataProvider().addAttributes(fields)
-    datalayer.updateFields()
-
-    with open(path_to_csv, "r") as file:
-        reader = csv.DictReader(file)
-        for row in reader:
-            origin = int(row["O"])
-            destination = int(row["D"])
-            tons = float(row["Ton"])
-
-            feature = QgsFeature()
-            feature.setAttributes([origin, destination, tons])
-
-            datalayer.dataProvider().addFeature(feature)
-
-    if not datalayer.isValid():
-        print("Open layer failed to load!")
-    else:
-        QgsProject.instance().addMapLayer(datalayer)
+    QgsProject.instance().removeAllMapLayers()
 
 
 @pytest.fixture
@@ -131,195 +100,4 @@ def coquimbo_project(qgis_iface, create_example) -> AequilibraEMenu:
     yield ae
     ae.run_close_project()
     qgis_iface.messageBar().messages = {0: [], 1: [], 2: [], 3: []}
-
-
-@pytest.fixture
-def create_links_with_matrix():
-    layer = QgsVectorLayer("Linestring?crs=epsg:4326", "lines", "memory")
-    if not layer.isValid():
-        print("lines layer failed to load!")
-    else:
-        field_id = QgsField("link_id", QVariant.Int)
-        matrix_ab = QgsField("matrix_ab", QVariant.Int)
-        matrix_ba = QgsField("matrix_ba", QVariant.Int)
-        matrix_tot = QgsField("matrix_tot", QVariant.Int)
-
-        layer.dataProvider().addAttributes([field_id, matrix_ab, matrix_ba, matrix_tot])
-        layer.updateFields()
-
-        lines = [
-            [QgsPointXY(1, 0), QgsPointXY(1, 1)],
-            [QgsPointXY(1, 0), QgsPointXY(0, 0)],
-            [QgsPointXY(0, 0), QgsPointXY(0, 1)],
-            [QgsPointXY(0, 1), QgsPointXY(1, 1)],
-            [QgsPointXY(0, 0), QgsPointXY(1, 1)],
-        ]
-
-        attributes = ([1, 2, 3, 4, 5], [50, 42, 17, 32, 19], [50, 63, 18, 32, 11], [100, 105, 35, 64, 30])
-
-        features = []
-        for i, (line, fid, ab, ba, tot) in enumerate(zip(lines, *attributes)):
-            feature = QgsFeature()
-            feature.setGeometry(QgsGeometry.fromPolylineXY(line))
-            feature.setAttributes([fid, ab, ba, tot])
-            features.append(feature)
-
-        layer.dataProvider().addFeatures(features)
-
-        QgsProject.instance().addMapLayer(layer)
-
-
-@pytest.fixture
-def run_assignment(ae_with_project):
-    from aequilibrae.paths import TrafficAssignment, TrafficClass
-
-    project = ae_with_project.project
-    project.network.build_graphs()
-
-    graph = project.network.graphs["c"]
-    graph.set_graph("free_flow_time")
-    graph.set_skimming(["free_flow_time", "distance"])
-    graph.set_blocked_centroid_flows(False)
-
-    demand = project.matrices.get_matrix("demand.aem")
-    demand.computational_view(["matrix"])
-
-    assigclass = TrafficClass("car", graph, demand)
-
-    assig = TrafficAssignment()
-
-    assig.set_classes([assigclass])
-    assig.set_vdf("BPR")
-    assig.set_vdf_parameters({"alpha": "b", "beta": "power"})
-    assig.set_capacity_field("capacity")
-    assig.set_time_field("free_flow_time")
-    assig.set_algorithm("bfw")
-    assig.max_iter = 5
-    assig.rgap_target = 0.01
-    assig.execute()
-
-    assig.save_results("assignment")
-    assig.save_skims("assignment", which_ones="all", format="omx")
-
-    return ae_with_project
-
-
-@pytest.fixture
-def create_polygons_layer(request):
-    layer = QgsVectorLayer("Polygon?crs=epsg:4326", "polygon", "memory")
-    if not layer.isValid():
-        print("Polygon layer failed to load!")
-    else:
-        field_id = QgsField("ID", QVariant.Int)
-        field_zone_id = QgsField("zone_id", QVariant.Int)
-        nickname = QgsField("name", QVariant.String)
-
-        layer.dataProvider().addAttributes([field_id, field_zone_id, nickname])
-        layer.updateFields()
-
-        polys = [
-            [
-                QgsPointXY(-71.2487, -29.8936),
-                QgsPointXY(-71.2487, -29.8895),
-                QgsPointXY(-71.2441, -29.8895),
-                QgsPointXY(-71.2441, -29.8936),
-                QgsPointXY(-71.2487, -29.8936),
-            ],
-            [
-                QgsPointXY(-71.2401, -29.8945),
-                QgsPointXY(-71.2401, -29.8928),
-                QgsPointXY(-71.2375, -29.8928),
-                QgsPointXY(-71.2375, -29.8945),
-                QgsPointXY(-71.2401, -29.8945),
-            ],
-            [
-                QgsPointXY(-71.2329, -29.8800),
-                QgsPointXY(-71.2329, -29.8758),
-                QgsPointXY(-71.2280, -29.8758),
-                QgsPointXY(-71.2280, -29.8800),
-                QgsPointXY(-71.2329, -29.8800),
-            ],
-        ]
-
-        attributes = (request.param, [None, None, None])
-
-        features = []
-        for i, (poly, zone_id, name) in enumerate(zip(polys, *attributes)):
-            feature = QgsFeature()
-            feature.setGeometry(QgsGeometry.fromPolygonXY([poly]))
-            feature.setAttributes([i + 1, zone_id, name])
-            features.append(feature)
-
-        layer.dataProvider().addFeatures(features)
-
-        QgsProject.instance().addMapLayer(layer)
-
-    return layer
-
-
-@pytest.fixture
-def load_sfalls_from_layer(folder_path, request):
-
-    fldr_pth = "test/data/SiouxFalls_project" if request.param == None else folder_path
-
-    if fldr_pth == folder_path:
-        if not exists(fldr_pth):
-            makedirs(fldr_pth)
-        copyfile("test/data/SiouxFalls_project/SiouxFalls.gpkg", f"{fldr_pth}/SiouxFalls.gpkg")
-
-    path_to_gpkg = f"{fldr_pth}/SiouxFalls.gpkg"
-
-    # append the layername part
-    gpkg_links_layer = path_to_gpkg + "|layername=links"
-    gpkg_nodes_layer = path_to_gpkg + "|layername=nodes"
-
-    linkslayer = QgsVectorLayer(gpkg_links_layer, "Links layer", "ogr")
-    nodeslayer = QgsVectorLayer(gpkg_nodes_layer, "Nodes layer", "ogr")
-
-    if not linkslayer.isValid():
-        print("Links layer failed to load!")
-    else:
-        QgsProject.instance().addMapLayer(linkslayer)
-
-    if not nodeslayer.isValid():
-        print("Nodes layer failed to load!")
-    else:
-        QgsProject.instance().addMapLayer(nodeslayer)
-        var = QgsProject.instance().mapLayersByName("Nodes layer")
-        if not var[0].crs().isValid():
-            crs = QgsCoordinateReferenceSystem("EPSG:4326")
-            var[0].setCrs(crs)
-
-
-@pytest.fixture
-def load_synthetic_future_vector():
-    import csv
-
-    path_to_csv = "test/data/SiouxFalls_project/synthetic_future_vector.csv"
-
-    datalayer = QgsVectorLayer("None?delimiter=,", "synthetic_future_vector", "memory")
-
-    fields = [
-        QgsField("index", QVariant.Int),
-        QgsField("origins", QVariant.Double),
-        QgsField("destinations", QVariant.Double),
-    ]
-    datalayer.dataProvider().addAttributes(fields)
-    datalayer.updateFields()
-
-    with open(path_to_csv, "r") as file:
-        reader = csv.DictReader(file)
-        for row in reader:
-            origin = float(row["origins"])
-            destination = float(row["destinations"])
-            index = int(row["index"])
-
-            feature = QgsFeature()
-            feature.setAttributes([index, origin, destination])
-
-            datalayer.dataProvider().addFeature(feature)
-
-    if not datalayer.isValid():
-        print("Open layer failed to load!")
-    else:
-        QgsProject.instance().addMapLayer(datalayer)
+    QgsProject.instance().removeAllMapLayers()
