+++ test/test_add_centroid_connector.py
@@ -1,16 +1,18 @@
 import pytest
 from shapely.geometry import Point
-from aequilibrae.utils.db_utils import read_and_close
-from aequilibrae.project.database_connection import database_connection
-from qgis.core import QgsProject, QgsVectorLayer, QgsField, QgsFeature
-from qgis.core import QgsPointXY, QgsGeometry
-from qgis.PyQt.QtCore import QVariant
 
+from aequilibrae.project.database_connection import database_connection
+from aequilibrae.utils.db_utils import read_and_close
+
+from qgis.core import QgsProject
+
+from .utilities import load_points_vector
 from qaequilibrae.modules.network.adds_connectors_dialog import AddConnectorsDialog
+from qaequilibrae.modules.network.adds_connectors_procedure import AddsConnectorsProcedure
 
 
 @pytest.mark.parametrize("in_zone", [True, False])
-def test_add_connectors_from_zones(pt_no_feed, in_zone):
+def test_add_connectors_from_zones_dialog(pt_no_feed, in_zone):
     dialog = AddConnectorsDialog(pt_no_feed)
     dialog.rdo_zone.setChecked(True)
     dialog.chb_zone.setChecked(in_zone)
@@ -28,8 +30,33 @@ def test_add_connectors_from_zones(pt_no_feed, in_zone):
         node_count = conn.execute("select count(node_id) from nodes where is_centroid=1").fetchone()[0]
         link_count = conn.execute("select count(name) from links where name like 'centroid connector%'").fetchone()[0]
 
-    assert node_count == 11
-    assert link_count == 11
+    assert node_count == 11 and link_count == 11
+
+
+def test_add_connectors_from_zones_procedure(pt_no_feed):
+    dialog = AddConnectorsDialog(pt_no_feed)
+    dialog.source = "zone"
+
+    zoning = dialog.project.zoning
+    assert len(zoning.all_zones())
+
+    parameters = {
+        "qgis_project": dialog.qgis_project,
+        "link_types": ["p"],
+        "modes": ["c"],
+        "num_connectors": 2,
+        "source": "zone",
+        "limit_to_zone": False,
+    }
+
+    dialog.worker_thread = AddsConnectorsProcedure(pt_no_feed.iface.mainWindow(), **parameters)
+    dialog.worker_thread.doWork()
+
+    with read_and_close(database_connection("network")) as conn:
+        node_count = conn.execute("select count(node_id) from nodes where is_centroid=1").fetchone()[0]
+        link_count = conn.execute("select count(name) from links where name like 'centroid connector%'").fetchone()[0]
+
+    assert node_count == 11 and link_count == 22
 
 
 @pytest.mark.parametrize(
@@ -39,7 +66,7 @@ def test_add_connectors_from_zones(pt_no_feed, in_zone):
         (101, 5555, Point(-71.3346, -29.9176)),
     ],
 )
-def test_add_connectors_from_network(pt_no_feed, node_id, radius, point):
+def test_add_connectors_from_network_dialog(pt_no_feed, node_id, radius, point):
     dialog = AddConnectorsDialog(pt_no_feed)
     dialog.rdo_network.setChecked(True)
 
@@ -63,43 +90,42 @@ def test_add_connectors_from_network(pt_no_feed, node_id, radius, point):
         link_count = conn.execute("select count(name) from links where name like 'centroid connector%'").fetchone()[0]
 
     if radius == 500:
-        assert node_count == 1
-        assert link_count == 0
+        assert node_count == 1 and link_count == 0
     else:
-        assert node_count == 1
-        assert link_count == 1
+        assert node_count == 1 and link_count == 1
 
 
-def test_add_connectors_from_layer(pt_no_feed):
-    nodes_layer = QgsVectorLayer("Point?crs=epsg:4326", "Centroids", "memory")
-    if not nodes_layer.isValid():
-        print("Nodes layer failed to load!")
-    else:
-        field_id = QgsField("ID", QVariant.Int)
-        nodes_layer.dataProvider().addAttributes([field_id])
+def test_add_connectors_from_network_procedure(
+    pt_no_feed,
+):
+    dialog = AddConnectorsDialog(pt_no_feed)
 
-        field_zone_id = QgsField("zone_id", QVariant.Int)
-        nodes_layer.dataProvider().addAttributes([field_zone_id])
+    nodes = dialog.project.network.nodes
+    nd = nodes.new_centroid(101)
+    nd.geometry = Point(-71.3346, -29.9176)
+    nd.save()
 
-        nodes_layer.updateFields()
-        points = [
-            QgsPointXY(-71.3509, -29.9393),
-            QgsPointXY(-71.3182, -29.9619),
-            QgsPointXY(12.4606, 41.9093),
-        ]
+    parameters = {
+        "qgis_project": dialog.qgis_project,
+        "link_types": ["p"],
+        "modes": ["c"],
+        "num_connectors": 1,
+        "source": "network",
+        "radius": 5555,
+    }
 
-        zone_ids = [97, 98, 99]
+    dialog.worker_thread = AddsConnectorsProcedure(pt_no_feed.iface.mainWindow(), **parameters)
+    dialog.worker_thread.doWork()
 
-        features = []
-        for i, (point, zone_id) in enumerate(zip(points, zone_ids)):
-            feature = QgsFeature()
-            feature.setGeometry(QgsGeometry.fromPointXY(point))
-            feature.setAttributes([i + 1, zone_id])
-            features.append(feature)
+    with read_and_close(database_connection("network")) as conn:
+        node_count = conn.execute("select count(node_id) from nodes where is_centroid=1").fetchone()[0]
+        link_count = conn.execute("select count(name) from links where name like 'centroid connector%'").fetchone()[0]
 
-        nodes_layer.dataProvider().addFeatures(features)
+    assert node_count == 1 and link_count == 1
 
-        QgsProject.instance().addMapLayer(nodes_layer)
+
+def test_add_connectors_from_layer_dialog(pt_no_feed):
+    load_points_vector()
 
     dialog = AddConnectorsDialog(pt_no_feed)
     dialog.rdo_layer.setChecked(True)
@@ -115,7 +141,32 @@ def test_add_connectors_from_layer(pt_no_feed):
         node_count = conn.execute("select count(node_id) from nodes where is_centroid=1").fetchone()[0]
         link_count = conn.execute("select count(name) from links where name like 'centroid connector%'").fetchone()[0]
 
-    assert node_count == 3
-    assert link_count == 2
+    assert node_count == 3 and link_count == 2
 
-    QgsProject.instance().removeMapLayer(nodes_layer.id())
+
+def test_add_connectors_from_layer_procedure(pt_no_feed):
+    load_points_vector()
+
+    nodes_layer = QgsProject.instance().mapLayersByName("Centroids")[0]
+
+    dialog = AddConnectorsDialog(pt_no_feed)
+
+    parameters = {
+        "qgis_project": dialog.qgis_project,
+        "link_types": ["p"],
+        "modes": ["c"],
+        "num_connectors": 1,
+        "source": "layer",
+        "radius": 3000,
+        "layer": nodes_layer,
+        "field": "ID",
+    }
+
+    dialog.worker_thread = AddsConnectorsProcedure(pt_no_feed.iface.mainWindow(), **parameters)
+    dialog.worker_thread.doWork()
+
+    with read_and_close(database_connection("network")) as conn:
+        node_count = conn.execute("select count(node_id) from nodes where is_centroid=1").fetchone()[0]
+        link_count = conn.execute("select count(name) from links where name like 'centroid connector%'").fetchone()[0]
+
+    assert node_count == 3 and link_count == 2
