+++ test/test_add_zones.py
@@ -1,12 +1,11 @@
-import pytest
 from time import sleep
-from qgis.core import QgsProject
 
+from .utilities import create_polygons_layer
 from qaequilibrae.modules.project_procedures.adds_zones_dialog import AddZonesDialog
 
 
-@pytest.mark.parametrize("create_polygons_layer", [[97, 98, 99]], indirect=True)
-def test_add_zones(pt_project, create_polygons_layer):
+def test_add_zones(pt_project):
+    _ = create_polygons_layer([97, 98, 99])
 
     dialog = AddZonesDialog(pt_project)
     dialog.chb_add_centroids.setChecked(True)
@@ -17,5 +16,3 @@ def test_add_zones(pt_project, create_polygons_layer):
     sleep(2)
 
     assert len(pt_project.project.zoning.all_zones()) == 3
-
-    QgsProject.instance().clear()
