+++ test/test_gis_desire_lines.py
@@ -1,6 +1,7 @@
 import pytest
 from PyQt5.QtCore import Qt
 
+from .utilities import load_sfalls_from_layer
 from qaequilibrae.modules.gis.desire_lines_dialog import DesireLinesDialog
 
 
@@ -20,8 +21,9 @@ def test_click_create_without_layers(ae_with_project, qtbot):
     ), "Level 1 error message is missing"
 
 
-@pytest.mark.parametrize("load_sfalls_from_layer", [None], indirect=True)
-def test_click_create_with_layers(ae_with_project, qtbot, timeoutDetector, load_sfalls_from_layer):
+def test_click_create_with_layers(ae_with_project, qtbot, timeoutDetector):
+    load_sfalls_from_layer(None)
+
     dialog = DesireLinesDialog(ae_with_project)
     dialog.show()
     qtbot.addWidget(dialog)
@@ -36,8 +38,9 @@ def test_click_create_with_layers(ae_with_project, qtbot, timeoutDetector, load_
     # test that something cool happened on the map?
 
 
-@pytest.mark.parametrize("load_sfalls_from_layer", [None], indirect=True)
-def test_click_create_with_layers_desired_selected(ae_with_project, qtbot, timeoutDetector, load_sfalls_from_layer):
+def test_click_create_with_layers_desired_selected(ae_with_project, qtbot, timeoutDetector):
+    load_sfalls_from_layer(None)
+
     dialog = DesireLinesDialog(ae_with_project)
     dialog.show()
     qtbot.addWidget(dialog)
@@ -54,8 +57,9 @@ def test_click_create_with_layers_desired_selected(ae_with_project, qtbot, timeo
 # Other than that, there isn't much error handling, so testing with wrong params triggers exceptions raising to the top
 # For example, one would expect something like this:
 @pytest.mark.skip(reason="Error handling implementation is required for this test")
-@pytest.mark.parametrize("load_sfalls_from_layer", [None], indirect=True)
-def test_click_create_with_layers_with_wrong_id_param(ae_with_project, qtbot, load_sfalls_from_layer):
+def test_click_create_with_layers_with_wrong_id_param(ae_with_project, qtbot):
+    load_sfalls_from_layer(None)
+
     dialog = DesireLinesDialog(ae_with_project)
     dialog.show()
     qtbot.addWidget(dialog)
