+++ test/test_qgis_environment.py
@@ -1,19 +1,13 @@
-import unittest
-
 from qgis.core import QgsProviderRegistry
-
 from .utilities import get_qgis_app
 
 
-class QGISTest(unittest.TestCase):
-    """Test the QGIS Environment"""
+def test_qgis_environment():
+    """QGIS environment has the expected providers"""
 
-    def setUp(self) -> None:
-        _ = get_qgis_app()
+    _ = get_qgis_app()
 
-    def test_qgis_environment(self):
-        """QGIS environment has the expected providers"""
+    prov_reg = QgsProviderRegistry.instance()
 
-        prov_reg = QgsProviderRegistry.instance()
-        self.assertGreater(len(prov_reg.providerList()), 20, "Missing too many providers")
-        self.assertIn("spatialite", prov_reg.providerList())
+    assert len(prov_reg.providerList()) > 20, "Missing too many providers"
+    assert "spatialite" in prov_reg.providerList()
